Add KeyExchange base class
