+++ src/Tmds.Ssh/ECDHKeyExchange.cs
@@ -9,15 +9,14 @@ using System.Numerics;
 namespace Tmds.Ssh;
 
 // ECDH Key Exchange: https://tools.ietf.org/html/rfc5656#section-4
-class ECDHKeyExchange : IKeyExchangeAlgorithm
+class ECDHKeyExchange : KeyExchange, IKeyExchangeAlgorithm
 {
     private readonly ECCurve _ecCurve;
-    private readonly HashAlgorithmName _hashAlgorithmName;
 
     public ECDHKeyExchange(ECCurve ecCurve, HashAlgorithmName hashAlgorithmName)
+        : base(hashAlgorithmName)
     {
         _ecCurve = ecCurve;
-        _hashAlgorithmName = hashAlgorithmName;
     }
 
     public async Task<KeyExchangeOutput> TryExchangeAsync(KeyExchangeContext context, IHostKeyVerification hostKeyVerification, Packet firstPacket, KeyExchangeInput input, ILogger logger, CancellationToken ct)
@@ -36,14 +35,7 @@ class ECDHKeyExchange : IKeyExchangeAlgorithm
         var ecdhReply = ParceEcdhReply(ecdhReplyMsg);
 
         // Verify received key is valid.
-        connectionInfo.ServerKey = new HostKey(ecdhReply.public_host_key);
-        await hostKeyVerification.VerifyAsync(connectionInfo, ct).ConfigureAwait(false);
-
-        var publicHostKey = PublicKey.CreateFromSshKey(ecdhReply.public_host_key);
-        if (publicHostKey is RsaPublicKey rsaPublicKey && rsaPublicKey.KeySize < input.MinimumRSAKeySize)
-        {
-            throw new ConnectFailedException(ConnectFailedReason.KeyExchangeFailed, $"Server RSA key size {rsaPublicKey.KeySize} is less than {input.MinimumRSAKeySize}.", connectionInfo);
-        }
+        var publicHostKey = await VerifyHostKeyAsync(hostKeyVerification, input, ecdhReply.public_host_key, ct).ConfigureAwait(false);
 
         // Compute shared secret.
         BigInteger sharedSecret;
@@ -60,22 +52,9 @@ class ECDHKeyExchange : IKeyExchangeAlgorithm
         byte[] exchangeHash = CalculateExchangeHash(sequencePool, input.ConnectionInfo, input.ClientKexInitMsg, input.ServerKexInitMsg, ecdhReply.public_host_key.Data, q_c, ecdhReply.q_s, sharedSecret);
 
         // Verify the server's signature.
-        if (!publicHostKey.VerifySignature(input.HostKeyAlgorithms, exchangeHash, ecdhReply.exchange_hash_signature))
-        {
-            throw new ConnectFailedException(ConnectFailedReason.KeyExchangeFailed, "Signature does not match host key.", connectionInfo);
-        }
+        VerifySignature(publicHostKey, input.HostKeyAlgorithms, exchangeHash, ecdhReply.exchange_hash_signature, connectionInfo);
 
-        byte[] sessionId = input.ConnectionInfo.SessionId ?? exchangeHash;
-        byte[] initialIVC2S = CalculateKey(sequencePool, sharedSecret, exchangeHash, (byte)'A', sessionId, input.InitialIVC2SLength);
-        byte[] initialIVS2C = CalculateKey(sequencePool, sharedSecret, exchangeHash, (byte)'B', sessionId, input.InitialIVS2CLength);
-        byte[] encryptionKeyC2S = CalculateKey(sequencePool, sharedSecret, exchangeHash, (byte)'C', sessionId, input.EncryptionKeyC2SLength);
-        byte[] encryptionKeyS2C = CalculateKey(sequencePool, sharedSecret, exchangeHash, (byte)'D', sessionId, input.EncryptionKeyS2CLength);
-        byte[] integrityKeyC2S = CalculateKey(sequencePool, sharedSecret, exchangeHash, (byte)'E', sessionId, input.IntegrityKeyC2SLength);
-        byte[] integrityKeyS2C = CalculateKey(sequencePool, sharedSecret, exchangeHash, (byte)'F', sessionId, input.IntegrityKeyS2CLength);
-
-        return new KeyExchangeOutput(exchangeHash,
-            initialIVS2C, encryptionKeyS2C, integrityKeyS2C,
-            initialIVC2S, encryptionKeyC2S, integrityKeyC2S);
+        return CalculateKeyExchangeOutput(input, sequencePool, sharedSecret, exchangeHash);
     }
 
     private byte[] CalculateExchangeHash(SequencePool sequencePool, SshConnectionInfo connectionInfo, ReadOnlyPacket clientKexInitMsg, ReadOnlyPacket serverKexInitMsg, byte[] public_host_key, ECPoint q_c, ECPoint q_s, BigInteger sharedSecret)
@@ -107,58 +86,6 @@ class ECDHKeyExchange : IKeyExchangeAlgorithm
             hash.AppendData(segment.Span);
         }
         return hash.GetHashAndReset();
-    }
-
-    private byte[] CalculateKey(SequencePool sequencePool, BigInteger sharedSecret, byte[] exchangeHash, byte c, byte[] sessionId, int keyLength)
-    {
-        // https://tools.ietf.org/html/rfc4253#section-7.2
-
-        byte[] key = new byte[keyLength];
-        int keyOffset = 0;
-
-        // HASH(K || H || c || session_id)
-        using Sequence sequence = sequencePool.RentSequence();
-        var writer = new SequenceWriter(sequence);
-        writer.WriteMPInt(sharedSecret);
-        writer.Write(exchangeHash);
-        writer.WriteByte(c);
-        writer.Write(sessionId);
-
-        using IncrementalHash hash = IncrementalHash.CreateHash(_hashAlgorithmName);
-        foreach (var segment in sequence.AsReadOnlySequence())
-        {
-            hash.AppendData(segment.Span);
-        }
-        byte[] K1 = hash.GetHashAndReset();
-        Append(key, K1, ref keyOffset);
-
-        while (keyOffset != key.Length)
-        {
-            sequence.Clear();
-
-            // K3 = HASH(K || H || K1 || K2)
-            writer = new SequenceWriter(sequence);
-            writer.WriteMPInt(sharedSecret);
-            writer.Write(exchangeHash);
-            writer.Write(key.AsSpan(0, keyOffset));
-
-            foreach (var segment in sequence.AsReadOnlySequence())
-            {
-                hash.AppendData(segment.Span);
-            }
-            byte[] Kn = hash.GetHashAndReset();
-
-            Append(key, Kn, ref keyOffset);
-        }
-
-        return key;
-
-        static void Append(byte[] key, byte[] append, ref int offset)
-        {
-            int available = Math.Min(append.Length, key.Length - offset);
-            append.AsSpan().Slice(0, available).CopyTo(key.AsSpan(offset));
-            offset += available;
-        }
     }
 
     private BigInteger DeriveSharedSecret(ECDiffieHellman ecdh, ECPoint q)
